renaming cvs/multi tranable class and file
