+++ src/client/cvsTranable.interface.ts
