+++ src/client/multiTranable.interface.ts
